refactor and fmt
